Add files via upload
In this simple project I created a variable that has more than one value. At the beginning, the user will choose a number from zero to two and will print the result, which is a name.

I hope that this simple code will benefit anyone.

I allow anyone to use it and modify it.